Add mobile verification functionality

+++ app/models/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import (Column, Integer,
+from sqlalchemy import (Column, BigInteger, Integer,
                         String, Boolean, DateTime)
 
 from sqlalchemy.ext.declarative import declarative_base
@@ -11,7 +11,7 @@ class User(Base):
     __tablename__ = 'mobileusers'
 
     id = Column(Integer, primary_key=True)
-    mobile_no = Column(Integer, unique=True, nullable=False)
+    mobile_no = Column(BigInteger, unique=True, nullable=False)
     otp = Column(Integer, nullable=False)
     is_verified = Column(Boolean, default=False)
     created_at = Column(DateTime, nullable=True)
